requirements.txt
Django
psycopg2
social-auth-app-django
pusher
django-celery
django-request